requirements.txt
langchain
langchain-community
langchain-openai
langchain-qdrant
langchain-ollama
qdrant-client
pypdf
ollama
python-dotenv
tqdm